requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.1
gunicorn==20.1.0
oauth2client==4.1.3
gcloud==0.18.3